Enable guest user creation & create static tables

+++ app/controllers/api/v1/authentication_controller.rb
@@ -1,8 +1,8 @@
 module API::V1
   class AuthenticationController < APIController
     attr_reader :user
+    before_action :find_user, only: [:login]
     before_action :verify_otp, only: [:login, :sign_up]
-    before_action :find_user, only: [:login, :sign_up]
     #region dependencies
       require 'jwt_auth_token'
       include AuthHelper
@@ -13,7 +13,7 @@ module API::V1
       user = User.new(user_params)
       if user.save
         deleteOtp(params[:phone])
-        response  = get_token_data({ id: user.id })
+        response  = get_token_data({ uuid: user.uuid, guest: user.guest? })
         render json: { sign_up: response }, status: :ok
       else
         render_error(user.errors) and return
@@ -24,27 +24,33 @@ module API::V1
     def login
       render_error( ['Phone not registered, please sign up']) and return if user.blank?
       deleteOtp(params[:phone])
-      response = get_token_data({ id: user.id })
+      response = get_token_data({ uuid: user.uuid, guest: user.guest })
       render json: { login: response }, status: :ok
     end
 
     private
 
+    def guest?
+      params[:guest].to_s == "true" || @user.guest?
+    end
+
     def verify_otp
-        render_error(['Invalid OTP']) and return unless Otp.verify(params[:phone], params[:otp])
+      return if guest?
+      render_error(['Invalid OTP']) and return unless Otp.verify(params[:phone], params[:otp])
     end
 
     def find_user
-      @user = User.find_by(phone: params[:phone])
+      @user = User.find_by(uuid: params[:uuid])
     end
 
     def deleteOtp(phone)
+      return if params[:guest].to_s == "true"
       otp = Otp.where(phone: phone)
       otp.destroy_all if otp.present?
     end
 
     def user_params
-      params.permit( :full_name, :phone)
+      params.permit( :full_name, :phone, :guest)
     end
   end
 end
